+++ src/components/search/SearchButtonBar.jsx
@@ -2,25 +2,33 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import SearchButton from './SearchButton';
 import SearchClearButton from './SearchClearButton';
-// import SearchSaveButton from './SearchSaveButton';
+import SearchPinButton from './SearchPinButton';
 import styles from '../../../styles/cspace-ui/ButtonBar.css';
+import PinnedQueriesButton from './PinnedQueriesButton';
 
 const propTypes = {
   onClearButtonClick: PropTypes.func,
+  onPinButtonClick: PropTypes.func,
+  onPinnedQueriesButtonClick: PropTypes.func,
 };
 
 const defaultProps = {
   onClearButtonClick: undefined,
+  onPinButtonClick: undefined,
+  onPinnedQueriesButtonClick: undefined,
 };
 
 export default function SearchButtonBar(props) {
   const {
     onClearButtonClick,
+    onPinButtonClick,
+    onPinnedQueriesButtonClick,
   } = props;
 
   return (
     <div className={styles.common}>
-      {/* <SearchSaveButton /> */}
+      {onPinnedQueriesButtonClick && <PinnedQueriesButton onClick={onPinnedQueriesButtonClick} />}
+      {onPinButtonClick && <SearchPinButton onClick={onPinButtonClick} />}
       <SearchButton />
       <SearchClearButton onClick={onClearButtonClick} />
     </div>